Move message helper functions to separate file
