+++ pkg/provisioningv2/rke2/planner/planner.go
@@ -7,7 +7,6 @@ import (
 	"math"
 	"path/filepath"
 	"reflect"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -352,46 +351,6 @@ func (p *Planner) Process(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlan
 	}
 
 	return status, nil
-}
-
-func atMostThree(names []string) string {
-	sort.Strings(names)
-	if len(names) > 3 {
-		return fmt.Sprintf("%s and %d more", strings.Join(names[:3], ","), len(names)-3)
-	}
-	return strings.Join(names, ",")
-}
-
-func detailedMessage(machines []string, messages map[string][]string) string {
-	if len(machines) != 1 {
-		return ""
-	}
-	message := messages[machines[0]]
-	if len(message) != 0 {
-		return fmt.Sprintf(": %s", strings.Join(message, ", "))
-	}
-	return ""
-}
-
-func removeReconciledCondition(machine *capi.Machine) *capi.Machine {
-	if machine == nil || len(machine.Status.Conditions) == 0 {
-		return machine
-	}
-
-	conds := make([]capi.Condition, 0, len(machine.Status.Conditions))
-	for _, c := range machine.Status.Conditions {
-		if string(c.Type) != string(rke2.Reconciled) {
-			conds = append(conds, c)
-		}
-	}
-
-	if len(conds) == len(machine.Status.Conditions) {
-		return machine
-	}
-
-	machine = machine.DeepCopy()
-	machine.SetConditions(conds)
-	return machine
 }
 
 // getControlPlaneJoinURL will return the first encountered join URL based on machine annotations for machines that are
